Run tasks, calculate resources

+++ db/index.sql
@@ -32,7 +32,7 @@ create table resources (
   id serial not null primary key,
   user_id int not null,
   resource text not null,
-  amount int not null,
+  amount double precision not null,
   -- Model Objects
   created_at timestamp without time zone not null default now(),
   updated_at timestamp without time zone
